first draft of flux-burst-local
this is an early prototype for a local bursting plugin.
We have a slurm variant that can grab variables from
slurm. This is working to generate config files,
and next we need to start the first broker in setup()
and then move onto an HPC system and test starting
workers

Signed-off-by: vsoch <vsoch@users.noreply.github.com>