get through steps of providing command for final burst
For the time being while we are developing, it is nice to have all the configs
generated and then a startscript generated we can manually run. This commit
adds all these steps, and next I need to move it onto an allocation and test
how to actually turn on the other hosts.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>